modules/ros_chatbot/requirements.txt
aiofiles
airtable-python-wrapper==0.15.2
boto3
bs4
chainmap==1.0.3
coloredlogs==14.0

emoji
fastapi
flask==0.10.1
gevent==20.6.2
google-cloud-dialogflow==2.0.0
google-cloud-translate==1.3.1
googlemaps==2.5.1
grpcio

##### system dependencies
# ros-noetic-ddynamic-reconfigure-python

langchain-anthropic
langchain-aws
langchain-community
langchain-openai>=0.1.25

lxml
mysqlclient==1.4.6
numpy==1.21.2

opencv-python==4.9.0.80
openpyxl==3.0.7
pandas==1.2.5

pydantic>=2.0,<3.0
pytest==4.6.11
python-benedict>=0.24.1
python-socketio~=5.5.2
PyYAML==5.3.1
rospkg~=1.2.3
sqlalchemy==1.3.18
telethon==1.21.1
transitions==0.9.0
uvicorn
websocket-client==0.57.0
